Fix RPM's init.d script to use `pidfilepath` for pidfile instead of `dbpath`.
The init.d script included in the RPM looks for `dbpath=...` in the conf file and sets that as $DBPATH, then does the same for $PIDFILE. However it seems from looking at the conf file that the init.d script should look for `pidfilepath = ` instead of `dbpath=`. This commit fixes the init.d script to look for the correct setting, so the pidfile is created in /var/run/mongodb/ as intended instead of in /var/lib/mongo as is today.